Add rudimentary power management, replace inefficient delays in idle loop with pin change interrupt.
Instead of wasting energy by idling via _delay_ms(), the CPU and
peripherals are now put to sleep and woken up by an external
(pin change) interrupt. After all peripherals are initialised, the pin
state is checked again (usually takes ~10ms) and an alarm is thrown if
the pin is HIGH.

This also changes the level of the pin monitoring, triggering on a
HIGH pin using the internal pull-up resistor instead of triggering
on a LOW pin, thus not requiring a pull-down resistor any longer.